Making models

+++ test/helper.js
@@ -3,15 +3,22 @@ global.sinon = require('sinon');
 global.proxyquire = require('proxyquire');
 global.tungus = require('tungus')
 global.mongoose = require('mongoose');
+global.Schema = mongoose.Schema;
+
 var fs = require('fs');
 var dbFolder = __dirname + '/db'
 
-before(function(){
+before(function(done){
   request = {};
   response = {};
 
   prepareDb(dbFolder);
   mongoose.connect('tingodb://'+dbFolder)
+
+  loadControllers();
+  setMockUp();
+
+  seedSample(done);
 })
 
 after(function(){
@@ -27,9 +34,64 @@ function prepareDb(folder){
 }
 
 function cleanDb(folder){
-  fs.readdirSync(folder).forEach(function(file,index){
-      var curPath = folder + "/" + file;
-      fs.unlinkSync(curPath);
-    });
-  fs.rmdir(folder)
+  if(fs.existsSync(folder)){
+    fs.readdirSync(folder).forEach(function(file,index){
+        var curPath = folder + "/" + file;
+        fs.unlinkSync(curPath);
+      });
+    fs.rmdir(folder)
+  }
+}
+
+function seedSample(done){
+  categoryMock = {
+      model: {
+        findOne: function(condition, callback){
+          callback({},category);
+        }
+      }
+    }
+  Color = require('../app/models/color');
+  Category = require('../app/models/category');
+  Site = proxyquire('../app/models/site', {
+    './category': categoryMock
+  });
+  seedColor(function(){
+    seedCategories(done);
+  })
+};
+
+function seedColor(done){
+  mongoose.model('Color').create({title: 'black', hex: '#000'}, function(err, created){
+    if(!err){
+      color = created;
+    }
+    done();
+  });
+}
+
+function seedCategories(done){
+  var params1 = {idName:'testcat1', title:'testcat1', color: color}
+  var params2 = {idName:'testcat2', title:'testcat2', color: color}
+  mongoose.model('Category').create([params1,params2], function(error, resultCategory){
+    category = resultCategory;
+    done();
+  })
+}
+
+function setMockUp(){
+  dataMock = {
+    sites: [
+      {name: 'test'},
+      {name: 'test2'}
+      ],
+    categories: [
+      {name: 'test3'},
+      {name: 'test4'}
+    ]
+  };
+}
+
+function loadControllers(done){
+  pagesController = require('../app/controllers/pages');
 }
